qtvcp -switch back to gobject for status signals
For what ever reason when using Qobject only the
signals from the last object would work.

The whole reason for gstat is so it's a global source for
all objects and passes signals to all of them.

Maybe this could be worked around but at the moment i wish to
move on in development.
It unfortunately means that in a handler file Gstat uses
Gobject style connections and all widgets use QObject style.

Signed-off-by: Chris Morley <chrisinnanaimo@hotmail.com>